Report correct unencrypted length
COSE_Encrypt_GetContent() returns a valid pointer to the content, but
the length parameter is set to 0.

The reason for this is because cbOut is set to the correct value, but is
then used as the remaining bytes in the EVP functions and gets set to 0.

This commit sets the output parameter to the correct value, so we can
use the payload safely.